Lift SearchModal out of CharacterUnit
