+++ components/CharacterGrid/index.tsx
@@ -1,5 +1,8 @@
 import React, { useState } from 'react'
+import { useModal as useModal } from '~utils/useModal'
+
 import CharacterUnit from '~components/CharacterUnit'
+import SearchModal from '~components/SearchModal'
 
 import './index.scss'
 
@@ -19,10 +22,29 @@ interface Props {
 }
 
 const CharacterGrid = (props: Props) => {
+    const { open, openModal, closeModal } = useModal()
+    const [searchPosition, setSearchPosition] = useState(0)
+
     const numCharacters: number = 5
+
+    function isCharacter(object: Character | Weapon | Summon): object is Character {
+        // There aren't really any unique fields here
+        return (object as Character).gender !== undefined
+    }
+
+    function openSearchModal(position: number) {
+        setSearchPosition(position)
+        openModal()
+    }
 
     function receiveCharacter(character: Character, position: number) {
         props.onSelect(GridType.Character, character, position)
+    }
+
+    function sendData(object: Character | Weapon | Summon, position: number) {
+        if (isCharacter(object)) {
+            receiveCharacter(object, position)
+        }
     }
 
     return (
@@ -33,8 +55,8 @@ const CharacterGrid = (props: Props) => {
                         return (
                             <li key={`grid_unit_${i}`} >
                                 <CharacterUnit 
+                                    onClick={() => { openSearchModal(i) }}
                                     editable={props.editable}
-                                    onReceiveData={receiveCharacter} 
                                     position={i} 
                                     character={props.grid[i]}
                                 />
@@ -42,6 +64,16 @@ const CharacterGrid = (props: Props) => {
                         )
                     })
                 }
+                {open ? (
+                <SearchModal 
+                    grid={props.grid}
+                    close={closeModal}
+                    send={sendData}
+                    fromPosition={searchPosition}
+                    object="characters"
+                    placeholderText="Search for a character..."
+                />
+            ) : null}
             </ul>
         </div>
     )
